Don't set a field if all of its components are nil
This partially restores the behavior introduced in 7974d48, but does so
in a way that does not omit nil components completely.

The original intent of this commit was to make it so that a vCard cannot
be created without a full name. Prior to this commit, the following
would hold true:

``` ruby
vcard = VCardigan.create
vcard.fn(nil)
vcard.valid?  #=> true
```

Now this is no longer the case:

``` ruby
vcard = VCardigan.create
vcard.fn(nil)
vcard.valid?  #=> false
```

(This behavior applies to other fields as well.)