VCard#respond_to? returns true for missing methods
Since VCard overrides #method_missing to always do something, it stands
to reason that #responds_to_missing? should always return true. This
makes it possible to decorate an instance of VCard using
SimpleDelegator, specifically so that sending missing methods to the
decorator to access or set fields will delegate onwards to the VCard
instance.